Style changes, public port
Adding a public port option. If a publicPort config variable is present,
specify this in the socket.io connection. If not present, assume port
80.

Also removed the code to set the user position based on who joined
first. Instead, the user always sees themselves on top. Planning to also
change the code to use a list intead of top and bottom attributes, so
that if an API is implemented and multiuser chat is supported, it's just
a matter of changing the frontend code.