Address Tickets 2, 7, 15
The currently-updating line in each scrollback div is now hilighted in
blue.
The input box takes focus whenever enableInput is called.
Adjust line-height a little.
Move the input box closer to the bottom of the top-box div. Using
absolute positioning, but may want to just do it via relative
positioning later.
Added a placeholder attribute. HTML5 only, but hey, we're living in the
future.
Adjust size and color of input box and notices.