Posterity
Input box is off page, kept in focus by keypress events on the page,
which also get sent to the input box. I have a clever little cursor as
well, but the lag is terrible, so I'm committing this just so I can get
it out of my memory as soon as possible.