Remove the textbox (but not really)
Instead of echoing output to the scrollback I'm just including the
textbox as part of the scrollback, and styling it to be unobtrusive.

Only issue so far is there is a (-1,-1) pixel shift when hitting enter
because the size of the input is slightly off from the line-height. I'll
fix that soon.